// src/app/model/campaign.ts
export class Campaign {
    campaignId: string;
    description: string;
    dateCreated: any;
    dateDeleted: string;
    campaignRating: string;
    campaignAddedBy: string;
    campaignLength: number;
    userId: string;
    title: string;
    status: number;
    uploaderBrand: string;
    digitalAgency: string;
    campaignRepositoryLink: string;
    campaignStatus: number;
    deletedDate: string;
    categoryList:null;
    contentStatus:null;
    deletedUserId: string;
}